Remove catch-up system...
...in favour of making everything* globally loaded instead.
This will allow the factory to keep going while away from the base. Initially, I was hesitant to make everything global so as to not cause a lot of lag with many components, but with my future plans, this is a required change and I want to catch bugs early.
*excludes the ItemInterface and BlueprintImprinter objects, as the interface does nothing except consume power on its own (all container logic is handled by BaseData) and BlueprintImprinter cannot be automated